--- sail/csrc/core/ops/loss/loss.h
@@ -1,2 +1,0 @@
-#include "cross_entropy.h"
-#include "mean_squared_error.h"
