+++ sail/csrc/core/ops/ops.h
@@ -3,16 +3,11 @@
 #include "CompareOps.h"
 #include "CopyOps.h"
 #include "ExponentialOps.h"
+#include "LinalgOps.h"
+#include "LossOps.h"
+#include "MutationOps.h"
 #include "ReductionOps.h"
 #include "ReprOps.h"
 #include "UnaryOps.h"
 
 #include "broadcast.h"
-#include "linalg/dots.h"
-#include "transformations/expand_dims.h"
-#include "transformations/reshape.h"
-#include "transformations/roll_axis.h"
-#include "transformations/squeeze.h"
-#include "transformations/transpose.h"
-
-#include "loss/loss.h"
