--- sail/csrc/core/ops/transformations/reshape.h
@@ -1,15 +1,0 @@
-#pragma once
-
-#include <iostream>
-
-#include "Tensor.h"
-#include "tensor_shape.h"
-
-namespace sail {
-
-namespace ops {
-
-Tensor reshape(const Tensor& tensor1, const TensorShape& new_shape);
-}  // namespace ops
-
-}  // namespace sail
